backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elasticsearch>=8.15.1",
    "fastapi[standard]>=0.115.4",
    "python-dotenv>=1.0.1",
]
